+++ pom.xml
@@ -205,6 +205,12 @@
 
             <dependency>
                 <groupId>cn.dynamictp</groupId>
+                <artifactId>dynamic-tp-spring-boot-starter-common</artifactId>
+                <version>${revision}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>cn.dynamictp</groupId>
                 <artifactId>dynamic-tp-spring-boot-starter-adapter-common</artifactId>
                 <version>${revision}</version>
             </dependency>
