+++ starter/starter-common/src/main/resources/META-INF/services/com.dtp.core.spring.PropertiesBinder
@@ -1,0 +1,1 @@
+com.dtp.starter.common.autoconfigure.spring.SpringBootBinder
